tx response parser refactor
