+++ framework/snippets/src/network_response.rs
