+++ framework/snippets/tests/test_tx_issued_token_identifier.rs
@@ -1,4 +1,4 @@
-use multiversx_sc_snippets::tx_response_from_network;
+use multiversx_sc_snippets::network_response;
 use multiversx_sdk::data::transaction::{TransactionInfo, TransactionOnNetwork};
 
 #[test]
@@ -203,7 +203,7 @@ fn test_process_issued_token_identifier_fungible() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Option<String> = Some("EGLDMEX-95c6d5".to_string());
 
@@ -362,7 +362,7 @@ fn test_process_issued_token_identifier_semi_fungible() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Option<String> = Some("DOPETEST-77200c".to_string());
 
@@ -618,7 +618,7 @@ fn test_process_issued_token_identifier_non_fungible() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Option<String> = Some("GEN-868593".to_string());
 
@@ -919,7 +919,7 @@ fn test_process_issued_token_identifier_meta_esdt() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Option<String> = Some("AVASH-7d8b5d".to_string());
 
@@ -1152,7 +1152,7 @@ fn test_set_special_roles_should_not_process_issued_token_identifier() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Option<String> = None;
 
@@ -1414,7 +1414,7 @@ fn test_multisig_issue_nft_and_set_all_roles() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected = Some("TESTCOLL1-5aa80c".to_string());
