+++ framework/snippets/tests/test_tx_multi_contract_sc_result.rs
@@ -1,4 +1,4 @@
-use multiversx_sc_snippets::tx_response_from_network;
+use multiversx_sc_snippets::network_response;
 use multiversx_sdk::data::transaction::{TransactionInfo, TransactionOnNetwork};
 
 #[test]
@@ -79,7 +79,7 @@ fn test_with_multi_contract_same_shard_tx_that_has_no_sc_result() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Vec<Vec<u8>> = vec![
         hex::decode("0a").unwrap(),
@@ -211,7 +211,7 @@ fn test_with_multi_contract_cross_shard_tx_that_has_no_callback() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Vec<Vec<u8>> = vec![];
 
@@ -340,7 +340,7 @@ fn test_with_multi_contract_cross_shard_tx_that_has_non_returning_callback() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Vec<Vec<u8>> = vec![];
 
@@ -469,7 +469,7 @@ fn test_with_multi_contract_cross_shard_tx_that_has_returning_callback() {
         .data
         .unwrap()
         .transaction;
-    let tx_response = tx_response_from_network::parse_tx_response(tx_on_network);
+    let tx_response = network_response::parse_tx_response(tx_on_network);
 
     let expected: Vec<Vec<u8>> = vec![];
